+++ src/Bridge/Symfony/Bundle/DependencyInjection/Configuration.php
@@ -120,6 +120,7 @@ final class Configuration implements ConfigurationInterface
                 ->booleanNode('enable_entrypoint')->defaultTrue()->info('Enable the entrypoint')->end()
                 ->booleanNode('enable_docs')->defaultTrue()->info('Enable the docs')->end()
                 ->booleanNode('enable_profiler')->defaultTrue()->info('Enable the data collector and the WebProfilerBundle integration.')->end()
+                ->booleanNode('enable_symfony_violation_normalizer')->defaultFalse()->info('Enable the Symfony ConstraintViolationListNormalizer.')->end()
                 ->arrayNode('collection')
                     ->addDefaultsIfNotSet()
                     ->children()