+++ src/Problem/Serializer/ConstraintViolationListNormalizer.php
@@ -13,8 +13,10 @@ declare(strict_types=1);
 
 namespace ApiPlatform\Core\Problem\Serializer;
 
-use ApiPlatform\Core\Serializer\AbstractConstraintViolationListNormalizer;
+use ApiPlatform\Core\Serializer\ConstraintViolationListNormalizerTrait;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
+use Symfony\Component\Serializer\Normalizer\ConstraintViolationListNormalizer as BaseConstraintViolationListNormalizer;
+use Symfony\Component\Validator\ConstraintViolationListInterface;
 
 /**
  * Converts {@see \Symfony\Component\Validator\ConstraintViolationListInterface} the API Problem spec (RFC 7807).
@@ -23,8 +25,10 @@ use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
  */
-final class ConstraintViolationListNormalizer extends AbstractConstraintViolationListNormalizer
+final class ConstraintViolationListNormalizer extends BaseConstraintViolationListNormalizer
 {
+    use ConstraintViolationListNormalizerTrait;
+
     public const FORMAT = 'jsonproblem';
     public const TYPE = 'type';
     public const TITLE = 'title';
@@ -34,11 +38,29 @@ final class ConstraintViolationListNormalizer extends AbstractConstraintViolatio
         self::TITLE => 'An error occurred',
     ];
 
-    public function __construct(array $serializePayloadFields = null, NameConverterInterface $nameConverter = null, array $defaultContext = [])
-    {
-        parent::__construct($serializePayloadFields, $nameConverter);
+    private $serializePayloadFields;
 
+    private $nameConverter;
+
+    private $useSymfonyNormalizer;
+
+    public function __construct(array $serializePayloadFields = null, NameConverterInterface $nameConverter = null, array $defaultContext = [], bool $useSymfonyNormalizer = false)
+    {
         $this->defaultContext = array_merge($this->defaultContext, $defaultContext);
+        $this->nameConverter = $nameConverter;
+
+        parent::__construct($this->defaultContext, $this->nameConverter);
+
+        $this->serializePayloadFields = $serializePayloadFields;
+        $this->useSymfonyNormalizer = $useSymfonyNormalizer;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsNormalization($data, $format = null): bool
+    {
+        return static::FORMAT === $format && $data instanceof ConstraintViolationListInterface && parent::supportsNormalization($data, $format);
     }
 
     /**
@@ -46,6 +68,13 @@ final class ConstraintViolationListNormalizer extends AbstractConstraintViolatio
      */
     public function normalize($object, $format = null, array $context = [])
     {
+        if ($this->useSymfonyNormalizer) {
+            ['type' => $type, 'title' => $title, 'detail' => $detail, 'violations' => $violations] = parent::normalize($object, $format, $context);
+            $this->addPayloadToViolations($object, $violations);
+
+            return ['type' => $type, 'title' => $title, 'detail' => $detail, 'violations' => $violations];
+        }
+
         [$messages, $violations] = $this->getMessagesAndViolations($object);
 
         return [